Clean up handler creation and server setup, set/keep idle task.
The handler now starts with a non-nil ended task meant to store the time of
creation. Also, the active task is only ended and not set to nil when it is
stopped/aborted. This way newly connected listeners can now be informed, how
long the server has been sitting idle.